local/requirements.txt
# Core ML/DL packages
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
gradio>=3.50.0

# Note: For Google Colab, these packages are automatically available
# For local PC, install using: pip install -r requirements.txt